requirements.txt
aiogram==3.13.1
kerykeion==4.26.3
python-dotenv==1.0.1
aiosqlite==0.20.0
